changes wrt pr comments
